se actualizo el coigo

+++ dbmanagerall/codigo.py
@@ -10,8 +10,6 @@ URL_API=os.environ.get("URL_API")
 maximo_dias_acumular=int(os.environ.get("DIAS_ACUMULAR"))
 connlocal = None
 cursorlocal=None
-# listaUsuariosServidor=[]
-# listaUsuariosLocal=[]
 listaHuellasServidor=[]
 listahuellaslocal=[]
 listaempleadosseguricel=[]
@@ -22,7 +20,7 @@ dias_acumulados=[]
 nroCaptahuellasConHuella=0
 nroCaptahuellasSinHuella=0
 captahuella_actual=0
-
+BorrarPeticionesListas=True
 ######################################
 #############CAPTAHUELLAS#############
 #######################################
@@ -98,6 +96,7 @@ while True:
                     if tablaCambiada == 'Usuarios':
                         try:
                             try:
+                                banderaUsuario=True
                                 cursorlocal.execute('SELECT cedula, nombre, telegram_id, internet, wifi, captahuella, rfid, facial FROM web_usuarios WHERE cedula=%s',(cedulaUsuario,))
                                 usuario_local= cursorlocal.fetchall()
 
@@ -123,6 +122,7 @@ while True:
                                                     nroCaptahuellasSinHuella=nroCaptahuellasSinHuella+1
                                                 except:
                                                     print(f"fallo al conectar con la esp8266 con la ip:{captahuella}")
+                                                    banderaUsuario=False
                                         if nroCaptahuellasSinHuella == captahuella_actual:
                                             cursorlocal.execute('DELETE FROM web_huellas WHERE id_suprema=%s', (id_suprema,))
                                             connlocal.commit()
@@ -131,7 +131,6 @@ while True:
                                         cursorlocal.execute('DELETE FROM web_usuarios WHERE cedula=%s', (cedulaUsuario,))
                                         cursorlocal.execute('DELETE FROM web_horariospermitidos WHERE cedula_id=%s', (cedulaUsuario,))
                                         connlocal.commit()
-                                    request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                                 elif not usuarioLocal and usuarioServidor: 
                                     for consultajson in request_json_usuario:
                                         cedula=consultajson['cedula']
@@ -145,7 +144,6 @@ while True:
                                     cursorlocal.execute('''INSERT INTO web_usuarios (cedula, nombre, telegram_id, internet, wifi, captahuella, rfid, facial)
                                     VALUES (%s, %s, %s, %s, %s, %s, %s, %s)''', (cedula, nombre, telegram_id, internet, wifi, captahuella, rfid, facial))
                                     connlocal.commit()
-                                    request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                                 elif usuarioLocal and usuarioServidor:
                                     for consultajson in request_json_usuario:
                                         cedula=consultajson['cedula']
@@ -157,14 +155,18 @@ while True:
                                         facial=consultajson['reconocimientoFacial']
                                     cursorlocal.execute("UPDATE web_usuarios SET telegram_id=%s, internet=%s, wifi=%s, captahuella=%s, rfid=%s, facial=%s WHERE cedula=%s", (telegram_id,internet,wifi,captahuella,rfid,facial,cedula))
                                     connlocal.commit()
-                                    request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                             except requests.exceptions.ConnectionError:
                                 print("fallo consultando api en usuarios")
+                                banderaUsuario=False
                         except Exception as e:
                             print(f"{e} - fallo total usuarios")
+                            banderaUsuario=False
+                        if banderaUsuario:
+                            requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                     elif tablaCambiada == 'Horarios':
                         try:
                             try:
+                                banderaHorario=True
                                 request_json_horarios = requests.get(url=f'{URL_API}obtenerhorariosindividualapi/{CONTRATO}/{cedulaUsuario}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3).json()
                                         
                                 horariosServidor=[]
@@ -195,11 +197,14 @@ while True:
                                         dia=horariosLocaliterar[3]
                                         cursorlocal.execute('DELETE FROM web_horariospermitidos WHERE entrada=%s AND salida=%s AND cedula_id=%s AND dia=%s',(entrada, salida, cedula, dia))
                                         connlocal.commit()
-                                request_json_horarios = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                             except requests.exceptions.ConnectionError:
-                                    print("fallo consultando api en horarios")
+                                print("fallo consultando api en horarios")
+                                banderaHorario=False
                         except Exception as e:
                             print(f"{e} - fallo total horarios")
+                            banderaHorario=False
+                        if banderaHorario:
+                            request_json_horarios = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                     elif tablaCambiada == 'Huellas':
                         try:
                             try:
@@ -285,15 +290,18 @@ while True:
                                                     requests.get(url=f'{captahuella}/quitar/{id_suprema_hex}', timeout=3)
                                                 except:
                                                     print(f"fallo al conectar con la esp8266 con la ip:{captahuella}")
-                                if banderaHuella:
-                                    requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                             except requests.exceptions.ConnectionError:
-                                    print("fallo consultando api en huellas")
+                                print("fallo consultando api en huellas")
+                                banderaHuella=False
                         except Exception as e:
                             print(f"{e} - fallo total huellas")
+                            banderaHuella=False
+                        if banderaHuella:
+                            requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                     elif tablaCambiada == 'Tags':
                         try:
                             try:
+                                banderaTag=True
                                 cursorlocal.execute('SELECT epc, cedula FROM web_tagsrfid WHERE cedula=%s', (cedulaUsuario,))
                                 tags_local= cursorlocal.fetchall()
                                 
@@ -320,13 +328,16 @@ while True:
                                         cedula=taglocaliterar[1]
                                         cursorlocal.execute('DELETE FROM web_tagsrfid WHERE epc=%s AND cedula=%s',(epc, cedula))
                                         connlocal.commit()
-                                request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                             except requests.exceptions.ConnectionError:
-                                    print("fallo consultando api en tags")
+                                print("fallo consultando api en tags")
+                                banderaTag=False
                         except Exception as e:
                             print(f"{e} - fallo total tags")
+                            banderaTag=False
+                        if banderaTag:
+                            requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                     else:
-                        request_json_usuario = requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
+                        requests.delete(url=f'{URL_API}eliminarcambioapi/{idCambio}/', auth=('BaseLocal_access', 'S3gur1c3l_local@'), timeout=3)
                 t1_cambios=tm.perf_counter()
             
             if total_log > TIEMPO_LOG:
@@ -371,9 +382,7 @@ while True:
                         if nro_int_local != nro_int_servidor:
 
                             for interaccion in interacciones_local:
-                                try:
-                                    listaLogsServidor.index(interaccion)
-                                except ValueError:
+                                if not interaccion in listaLogsServidor:
                                     nombre=interaccion[0]
                                     fecha=interaccion[1]
                                     hora=interaccion[2]
